Implemented pressure sensor code, it works but still needs to be fine tuned.
Added working notification bar when ph or water level is out of the threshold.

Added a graph view alongside the table view in the sensor_history.html.

Fixed TUS logo not working in settings.html also added a functional refill demo gauge that flashes an LED on the pi as its refilling.